removed bedrock-format.ts, added smal function for message formatting in sapaicore, removed models lazy loading, simplifying the code
